refactor: fold strategy failure into features
