+++ bot_trade/config/rl_callbacks.py
@@ -29,9 +29,8 @@ import csv
 import shutil
 
 from bot_trade.config.rl_paths import stamp_name, _atomic_replace
-from . import strategy_failure
+from . import strategy_features as strategy_failure
 from collections import Counter
-from pathlib import Path
 
 # Optional dependencies — كلّها اختيارية
 try:  # TensorBoard
