+++ bot_trade/config/strategy_features.py
@@ -8,10 +8,16 @@
 import os
 import yaml
 import logging
+import datetime as dt
+from pathlib import Path
+from typing import Any, Dict, List, Optional
+
 import numpy as np
 import pandas as pd
 import ta
 from ta.momentum import AwesomeOscillatorIndicator
+
+from bot_trade.tools.atomic_io import append_jsonl
 
 try:
     CFG_PATH = os.path.join("config", "config.yaml")
@@ -202,3 +208,142 @@ def add_strategy_features(df: pd.DataFrame) -> pd.DataFrame:
 
 
     return df.reset_index(drop=True)
+
+
+# ----------------------------------------------------------------------
+# Strategy failure policy (merged from strategy_failure.py)
+# ----------------------------------------------------------------------
+
+SF_CONFIG: Dict[str, Any] = {}
+SF_STATE: Dict[str, int] = {"level": 0, "cool": 0}
+
+
+def configure(cfg: Dict[str, Any]) -> None:
+    """Set global configuration for strategy failure policy."""
+    SF_CONFIG.clear()
+    SF_CONFIG.update(cfg or {})
+
+
+def _now() -> str:
+    return dt.datetime.utcnow().isoformat()
+
+
+def evaluate_step(ctx: Dict[str, Any]) -> List[Dict[str, Any]]:
+    """Return zero or more failure events detected this step."""
+    if not SF_CONFIG.get("enabled", False):
+        return []
+    thr = SF_CONFIG.get("thresholds", {}) or {}
+    events: List[Dict[str, Any]] = []
+    ts = ctx.get("ts") or _now()
+    for key, limit in thr.items():
+        val = ctx.get(key)
+        if val is None or limit is None:
+            continue
+        try:
+            v = float(val)
+            lim = float(limit)
+        except Exception:
+            continue
+        trig = v >= lim if key in {"loss_streak", "stuck_position_s", "partial_fill_timeout_s"} else v > lim
+        if trig:
+            events.append({
+                "flag": key,
+                "reason": f"{key} threshold",
+                "value": v,
+                "threshold": lim,
+                "ts": ts,
+            })
+    return events
+
+
+def _clamp(key: str, value: float) -> float:
+    clamp_cfg = (SF_CONFIG.get("clamps") or {}).get(key, {})
+    lo = float(clamp_cfg.get("min", float("-inf")))
+    hi = float(clamp_cfg.get("max", float("inf")))
+    return max(lo, min(value, hi))
+
+
+def apply_actions(
+    events: List[Dict[str, Any]],
+    *,
+    risk_manager: Any = None,
+    controller: Any = None,
+    env: Any = None,
+    log_path: Optional[Path] = None,
+) -> Dict[str, Any]:
+    """Escalate actions based on policy and record events."""
+    summary: Dict[str, Any] = {"applied_actions": [], "new_risk_bounds": {}}
+    if not events:
+        if SF_STATE.get("cool", 0) > 0:
+            SF_STATE["cool"] -= 1
+            if SF_STATE["cool"] <= 0:
+                SF_STATE["level"] = 0
+        return summary
+
+    SF_STATE["cool"] = int(SF_CONFIG.get("cool_down_steps", 0))
+    actions = SF_CONFIG.get("actions", [])
+    level = SF_STATE.get("level", 0)
+
+    for ev in events:
+        try:
+            if risk_manager is not None:
+                risk_manager.record_flag(ev["flag"], ev["reason"], ev["value"], ev["threshold"])
+        except Exception:
+            pass
+
+    applied: List[str] = []
+    new_bounds: Dict[str, float] = {}
+    if level < len(actions):
+        act = actions[level]
+        if act == "reduce_risk" and risk_manager is not None:
+            try:
+                factor = 0.5
+                if getattr(controller, "last_regime", None) in {"high_vol", "low_liquidity"}:
+                    factor *= 0.8
+                risk_manager.current_risk = _clamp("risk_scale", risk_manager.current_risk * factor)
+                new_bounds["risk_scale"] = risk_manager.current_risk
+                if env is not None and hasattr(env, "exec_sim"):
+                    env.exec_sim.max_spread_bp = _clamp(
+                        "max_spread_bp", getattr(env.exec_sim, "max_spread_bp", 0.0) * factor
+                    )
+                    new_bounds["max_spread_bp"] = env.exec_sim.max_spread_bp
+            except Exception:
+                pass
+        elif act == "freeze_trading" and env is not None:
+            try:
+                setattr(env, "trading_frozen", True)
+            except Exception:
+                pass
+        elif act == "flat_all" and env is not None:
+            try:
+                flat = getattr(env, "flat_all", None)
+                if callable(flat):
+                    flat()
+            except Exception:
+                pass
+        elif act == "halt_training":
+            summary["halt"] = True
+        applied.append(act)
+        level += 1
+        if controller is not None and getattr(controller, "log_path", None) and act != "warn":
+            record = {"ts": _now(), "source": "safety", "action": act, "risk_bounds": new_bounds}
+            try:
+                append_jsonl(controller.log_path, record)
+            except Exception:
+                pass
+
+    SF_STATE["level"] = level
+
+    if log_path is not None:
+        for ev in events:
+            rec = dict(ev)
+            rec["actions"] = applied
+            try:
+                append_jsonl(log_path, rec)
+            except Exception:
+                pass
+
+    summary["applied_actions"] = applied
+    summary["new_risk_bounds"] = new_bounds
+    return summary
+
