--- bot_trade/config/stratigy_failure.py
@@ -1,2 +1,0 @@
-from .strategy_failure import *  # noqa
-print("[DEPRECATED] bot_trade/config/stratigy_failure.py → use config/strategy_failure.py")
